Pindah logic activity
